move IntroText and its scss file to separate component folder
