+++ client/app/page.tsx
@@ -1,5 +1,5 @@
 import ChameleonGraphic from '@components/ChameleonGraphic'
-import IntroText from '@components/IntroText'
+import IntroText from '@components/IntroText/IntroText'
 import StartSwiperAndLinks from '@components/StartSwiperAndLinks'
 import styles from '@styles/App.module.scss'
