+++ client/components/EmailConfirmed/EmailConfirmed.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+function EmailConfirmed() {
+  return (
+    <div>
+      <h2>EmailConfirmed</h2>
+    </div>
+  )
+}
+
+export default EmailConfirmed
